update fitur transaksi peminjaman & pengembalian

+++ database/migrations/2026_05_12_234908_drop_tanggal_kembali_from_peminjaman.php
@@ -4,15 +4,20 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration {
+return new class extends Migration
+{
     /**
      * Run the migrations.
      */
-    public function up()
+    public function up(): void
     {
-        Schema::table('peminjaman', function (Blueprint $table) {
-            $table->dropColumn('tanggal_kembali');
-        });
+        if (Schema::hasColumn('peminjaman', 'tanggal_kembali')) {
+
+            Schema::table('peminjaman', function (Blueprint $table) {
+                $table->dropColumn('tanggal_kembali');
+            });
+
+        }
     }
 
     /**
@@ -20,8 +25,6 @@ return new class extends Migration {
      */
     public function down(): void
     {
-        Schema::table('peminjaman', function (Blueprint $table) {
-            //
-        });
+        //
     }
 };
